requirements.txt
streamlit
python-dotenv
openai
streamlit-lottie
langchain_openai
pinecone